improve:PPT联动
